+++ src/Starcounter.Internal/Profiler/Profiler.cs
@@ -239,7 +239,7 @@ namespace Starcounter
                     Byte schedId = i;
 
                     // Running asynchronous task.
-                    ScSessionClass.DbSession.RunAsync(() => {
+                    Scheduling.ScheduleTask(() => {
 
                         Byte currentSchedId = StarcounterEnvironment.CurrentSchedulerId;
                         allSchedProfilerResults[currentSchedId] = Profiler.Current.GetResultsInJson(false);
@@ -289,7 +289,7 @@ namespace Starcounter
                     Byte schedId = i;
 
                     // Running asynchronous task.
-                    ScSessionClass.DbSession.RunAsync(() => {
+                    Scheduling.ScheduleTask(() => {
 
                         Profiler.Current.ResetAll();
                         bmx.sc_profiler_reset(StarcounterEnvironment.CurrentSchedulerId);